add --ai flag support

+++ test_generator/test_handlers/vedro_handler.py
@@ -19,7 +19,8 @@ class VedroHandler(TestHandler):
     def read_test(self, file_path: str, *args, **kwargs) -> TestScenario:
         raise NotImplementedError('method is not implemented')
 
-    def write_test(self, file_path: str, scenario: TestScenario, feature: str, story: str, *args, **kwargs) -> None:
+    def write_test(self, file_path: str, scenario: TestScenario, feature: str, story: str,
+                   force: bool = False, *args, **kwargs) -> None:
         filled_template = self.template.replace('$feature', feature) \
                                        .replace('$story', story) \
                                        .replace('$priority', scenario.priority) \
@@ -27,19 +28,32 @@ class VedroHandler(TestHandler):
                                        .replace('$description', scenario.description) \
                                        .replace('$expected_result', scenario.expected_result) \
                                        .replace('$params', self.__get_params(scenario))
+        if not force and os.path.exists(file_path):
+            print(f"File already exists: {file_path}")
+            return
 
         with open(file_path, 'w', encoding='utf-8') as file:
             file.write(filled_template)
 
         print(f"Test file created: {file_path}")
 
-    def write_tests(self, dir_path: str, suite: Suite, *args, **kwargs) -> None:
+    def write_tests(self, dir_path: str, suite: Suite, force: bool = False, *args, **kwargs) -> None:
         if not os.path.exists(dir_path):
             os.makedirs(dir_path)
 
         for scenario in suite.test_scenarios:
-            test_path = os.path.join(dir_path, scenario.test_name)
-            self.write_test(file_path=test_path, scenario=scenario, feature=suite.feature, story=suite.story)
+            test_path = os.path.join(dir_path, self.get_file_name(scenario))
+            self.write_test(
+                file_path=test_path,
+                scenario=scenario,
+                feature=suite.feature,
+                story=suite.story,
+                force=force
+            )
+
+    def get_file_name(self, scenario: TestScenario) -> str:
+        file_name = scenario.test_name or f"{scenario.subject.strip().replace(' ', '_').replace('-', '_').lower()}.py"
+        return file_name
 
     def validate_suite(self, suite: Suite, *args, **kwargs) -> None:
         if not suite.feature:
@@ -53,8 +67,6 @@ class VedroHandler(TestHandler):
                 raise ScenariosValidationError(f'Priority is not defined for scnenario {scenario}')
             if not scenario.subject:
                 raise ScenariosValidationError(f'Subject is not defined for scnenario {scenario}')
-            if not scenario.test_name:
-                raise ScenariosValidationError(f'Test name is not defined for scnenario {scenario}')
 
     def __get_subject(self, scneario: TestScenario) -> str:
         append_str = " (param = {{param}})"
